chore: Update pubspec.yaml and enhance UI components
This commit updates the `pubspec.yaml` file to change the app description, update the Dart SDK version, and modify dependencies for improved functionality. Additionally, it refines the UI in `main.dart` and `home_screen.dart` by replacing button structures with `ElevatedButton.icon` and `TextButton.icon` for better consistency and accessibility. The `DiffViewWidget` and `YamlTreeView` have also been adjusted to improve layout and user experience. These changes aim to enhance the overall performance and visual appeal of the application.